third_party/WebKit/LayoutTests/plugins/npruntime/delete-plugin-within-hasProperty-return-true-expected.txt

CONSOLE ERROR: line 13: Uncaught ReferenceError: NPObject deleted
